layout/forms/test/test_unstyled_control_height.html

<!doctype html>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<style>
  #container * {
    white-space: nowrap;
    appearance: none;
  }
  input[type=date] {
    /* date by default uses a monospace font, which might have different metrics */
    font: -moz-field;
  }
  .small-font * {
    font-size: 8px !important; /* important to override rule above */
  }
  .no-padding * {
    padding: 0;
  }
</style>
<div id="container">
  <input>
  <input type=date>
  <button>Foo</button>
  <select><option>Foo</option></select>
</div>
<script>
function testHeightMatches(desc) {
  let commonHeight = null;
  let commonTop = null;
  for (let element of document.querySelectorAll("#container > *")) {
    let rect = element.getBoundingClientRect();
    if (commonHeight === null) {
      commonHeight = rect.height;
      commonTop = rect.top;
    }
    is(rect.height, commonHeight, `Height of the controls should match for ${element.outerHTML}${desc}`);
    is(rect.top, commonTop, `Top of the controls should match for ${element.outerHTML}${desc}`);
  }
}

const container = document.getElementById("container");

testHeightMatches("");

container.className = "no-padding";

testHeightMatches(" without padding");

container.className = "small-font";

testHeightMatches(" with an small font");
</script>
